feat: HTML element creator
